findById 매개변수 타입 수정

+++ jjikeobang-server/src/main/java/com/jjikeobang/member/repository/MemberRepository.java
@@ -16,7 +16,7 @@ public interface MemberRepository {
 
     void insertMember(JoinMemberDTO member);
 
-    Member findById(int memberId);
+    Member findById(Long memberId);
 
     Member findByLoginId(String userId);
 }
